+++ src/Symfony/Component/Validator/Tests/ConstraintValidatorTest.php
@@ -14,7 +14,9 @@ namespace Symfony\Component\Validator\Tests;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
-use Symfony\Component\Validator\Tests\Fixtures\TestEnum;
+use Symfony\Component\Validator\Tests\Fixtures\TestEnumBackendInteger;
+use Symfony\Component\Validator\Tests\Fixtures\TestEnumBackendString;
+use Symfony\Component\Validator\Tests\Fixtures\TestEnumUnit;
 
 class ConstraintValidatorTest extends TestCase
 {
@@ -49,7 +51,9 @@ class ConstraintValidatorTest extends TestCase
             [class_exists(\IntlDateFormatter::class) ? static::normalizeIcuSpaces("Feb 2, 1971, 8:00\u{202F}AM") : '1971-02-02 08:00:00', $dateTime, ConstraintValidator::PRETTY_DATE],
             [class_exists(\IntlDateFormatter::class) ? static::normalizeIcuSpaces("Jan 1, 1970, 6:00\u{202F}AM") : '1970-01-01 06:00:00', new \DateTimeImmutable('1970-01-01T06:00:00Z'), ConstraintValidator::PRETTY_DATE],
             [class_exists(\IntlDateFormatter::class) ? static::normalizeIcuSpaces("Jan 1, 1970, 3:00\u{202F}PM") : '1970-01-01 15:00:00', (new \DateTimeImmutable('1970-01-01T23:00:00'))->setTimezone(new \DateTimeZone('America/New_York')), ConstraintValidator::PRETTY_DATE],
-            ['FirstCase', TestEnum::FirstCase],
+            ['"FirstCase"', TestEnumUnit::FirstCase],
+            ['"a"', TestEnumBackendString::FirstCase],
+            ['3', TestEnumBackendInteger::FirstCase],
         ];
 
         date_default_timezone_set($defaultTimezone);