+++ src/Symfony/Component/Validator/CHANGELOG.md
@@ -56,6 +56,7 @@ CHANGELOG
    ```
  * Add support for ratio checks for SVG files to the `Image` constraint
  * Add the `Slug` constraint
+ * Add support for enums in `Choice` constraint
 
 7.2
 ---